# perf: optimize bind var generation