--- a/go/vt/sqlparser/ast_format_fast.go
+++ b/go/vt/sqlparser/ast_format_fast.go
@@ -1293,7 +1293,7 @@ func (node *Literal) formatFast(buf *TrackedBuffer) {
 
 // formatFast formats the node.
 func (node Argument) formatFast(buf *TrackedBuffer) {
-	buf.WriteArg(string(node))
+	buf.WriteArg(":", string(node))
 }
 
 // formatFast formats the node.
@@ -1342,7 +1342,7 @@ func (node *DerivedTable) formatFast(buf *TrackedBuffer) {
 
 // formatFast formats the node.
 func (node ListArg) formatFast(buf *TrackedBuffer) {
-	buf.WriteArg(string(node))
+	buf.WriteArg("::", string(node))
 }
 
 // formatFast formats the node.