--- a/go/vt/sqlparser/ast_funcs.go
+++ b/go/vt/sqlparser/ast_funcs.go
@@ -1296,11 +1296,11 @@ func (lock LockOptionType) ToString() string {
 }
 
 // CompliantName is used to get the name of the bind variable to use for this column name
-func (node *ColName) CompliantName(suffix string) string {
+func (node *ColName) CompliantName() string {
 	if !node.Qualifier.IsEmpty() {
-		return node.Qualifier.Name.CompliantName() + "_" + node.Name.CompliantName() + suffix
+		return node.Qualifier.Name.CompliantName() + "_" + node.Name.CompliantName()
 	}
-	return node.Name.CompliantName() + suffix
+	return node.Name.CompliantName()
 }
 
 // AtCount represents the '@' count in ColIdent