--- a/go/vt/sqlparser/ast_rewriting.go
+++ b/go/vt/sqlparser/ast_rewriting.go
@@ -17,6 +17,8 @@ limitations under the License.
 package sqlparser
 
 import (
+	"strconv"
+
 	querypb "vitess.io/vitess/go/vt/proto/query"
 	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
 	"vitess.io/vitess/go/vt/vterrors"
@@ -32,10 +34,109 @@ type RewriteASTResult struct {
 	AST Statement // The rewritten AST
 }
 
+type ReservedVars struct {
+	prefix       string
+	reserved     BindVars
+	next         []byte
+	counter      int
+	fast, static bool
+}
+
+func (r *ReservedVars) ContainsAny(names ...string) bool {
+	for _, name := range names {
+		if _, ok := r.reserved[name]; ok {
+			return true
+		}
+	}
+	for _, name := range names {
+		r.reserved[name] = struct{}{}
+	}
+	return false
+}
+
+func (r *ReservedVars) NextColumnName(col *ColName) string {
+	compliantName := col.CompliantName()
+	if r.fast && strings.HasPrefix(compliantName, r.prefix) {
+		compliantName = "_" + compliantName
+	}
+
+	joinVar := []byte(compliantName)
+	baseLen := len(joinVar)
+	i := int64(1)
+
+	for {
+		if _, ok := r.reserved[string(joinVar)]; !ok {
+			return string(joinVar)
+		}
+		joinVar = strconv.AppendInt(joinVar[:baseLen], i, 10)
+		i++
+	}
+}
+
+const staticBvar10 = "vtg0vtg1vtg2vtg3vtg4vtg5vtg6vtg7vtg8vtg9"
+const staticBvar100 = "vtg10vtg11vtg12vtg13vtg14vtg15vtg16vtg17vtg18vtg19vtg20vtg21vtg22vtg23vtg24vtg25vtg26vtg27vtg28vtg29vtg30vtg31vtg32vtg33vtg34vtg35vtg36vtg37vtg38vtg39vtg40vtg41vtg42vtg43vtg44vtg45vtg46vtg47vtg48vtg49vtg50vtg51vtg52vtg53vtg54vtg55vtg56vtg57vtg58vtg59vtg60vtg61vtg62vtg63vtg64vtg65vtg66vtg67vtg68vtg69vtg70vtg71vtg72vtg73vtg74vtg75vtg76vtg77vtg78vtg79vtg80vtg81vtg82vtg83vtg84vtg85vtg86vtg87vtg88vtg89vtg90vtg91vtg92vtg93vtg94vtg95vtg96vtg97vtg98vtg99"
+
+func (r *ReservedVars) Next() string {
+	if r.fast {
+		r.counter++
+
+		if r.static {
+			switch {
+			case r.counter < 10:
+				ofs := r.counter * 4
+				return staticBvar10[ofs : ofs+4]
+			case r.counter < 100:
+				ofs := (r.counter - 10) * 5
+				return staticBvar100[ofs : ofs+5]
+			}
+		}
+
+		r.next = strconv.AppendInt(r.next[:len(r.prefix)], int64(r.counter), 10)
+		return string(r.next)
+	}
+
+	for {
+		r.counter++
+		r.next = strconv.AppendInt(r.next[:len(r.prefix)], int64(r.counter), 10)
+
+		if _, ok := r.reserved[string(r.next)]; !ok {
+			bvar := string(r.next)
+			r.reserved[bvar] = struct{}{}
+			return bvar
+		}
+	}
+}
+
+func NewReservedVars(prefix string, known BindVars) *ReservedVars {
+	rv := &ReservedVars{
+		prefix:   prefix,
+		counter:  0,
+		reserved: known,
+		fast:     true,
+		next:     []byte(prefix),
+	}
+
+	if prefix != "" && prefix[0] == '_' {
+		panic("cannot reserve variables with a '_' prefix")
+	}
+
+	for bvar := range known {
+		if strings.HasPrefix(bvar, prefix) {
+			rv.fast = false
+			break
+		}
+	}
+
+	if prefix == "vtg" {
+		rv.static = true
+	}
+	return rv
+}
+
 // PrepareAST will normalize the query
-func PrepareAST(in Statement, reservedVars BindVars, bindVars map[string]*querypb.BindVariable, prefix string, parameterize bool, keyspace string) (*RewriteASTResult, error) {
+func PrepareAST(in Statement, reservedVars *ReservedVars, bindVars map[string]*querypb.BindVariable, parameterize bool, keyspace string) (*RewriteASTResult, error) {
 	if parameterize {
-		err := Normalize(in, reservedVars, bindVars, prefix)
+		err := Normalize(in, reservedVars, bindVars)
 		if err != nil {
 			return nil, err
 		}
@@ -323,7 +424,7 @@ func (er *expressionRewriter) unnestSubQueries(cursor *Cursor, subquery *Subquer
 }
 
 func bindVarExpression(name string) Expr {
-	return NewArgument(":" + name)
+	return NewArgument(name)
 }
 
 // SystemSchema returns true if the schema passed is system schema