--- a/go/vt/sqlparser/normalizer.go
+++ b/go/vt/sqlparser/normalizer.go
@@ -17,8 +17,6 @@ limitations under the License.
 package sqlparser
 
 import (
-	"strconv"
-
 	"vitess.io/vitess/go/sqltypes"
 
 	querypb "vitess.io/vitess/go/vt/proto/query"
@@ -34,27 +32,23 @@ type BindVars map[string]struct{}
 // Within Select constructs, bind vars are deduped. This allows
 // us to identify vindex equality. Otherwise, every value is
 // treated as distinct.
-func Normalize(stmt Statement, known BindVars, bindVars map[string]*querypb.BindVariable, prefix string) error {
-	nz := newNormalizer(known, bindVars, prefix)
+func Normalize(stmt Statement, reserved *ReservedVars, bindVars map[string]*querypb.BindVariable) error {
+	nz := newNormalizer(reserved, bindVars)
 	_ = Rewrite(stmt, nz.WalkStatement, nil)
 	return nz.err
 }
 
 type normalizer struct {
 	bindVars map[string]*querypb.BindVariable
-	prefix   string
-	reserved map[string]struct{}
-	counter  int
+	reserved *ReservedVars
 	vals     map[string]string
 	err      error
 }
 
-func newNormalizer(reserved map[string]struct{}, bindVars map[string]*querypb.BindVariable, prefix string) *normalizer {
+func newNormalizer(reserved *ReservedVars, bindVars map[string]*querypb.BindVariable) *normalizer {
 	return &normalizer{
 		bindVars: bindVars,
-		prefix:   prefix,
 		reserved: reserved,
-		counter:  1,
 		vals:     make(map[string]string),
 	}
 }
@@ -135,7 +129,7 @@ func (nz *normalizer) convertLiteralDedup(node *Literal, cursor *Cursor) {
 	bvname, ok := nz.vals[key]
 	if !ok {
 		// If there's no such bindvar, make a new one.
-		bvname = nz.newName()
+		bvname = nz.reserved.Next()
 		nz.vals[key] = bvname
 		nz.bindVars[bvname] = bval
 	}
@@ -151,7 +145,7 @@ func (nz *normalizer) convertLiteral(node *Literal, cursor *Cursor) {
 		return
 	}
 
-	bvname := nz.newName()
+	bvname := nz.reserved.Next()
 	nz.bindVars[bvname] = bval
 
 	cursor.Replace(NewArgument(":" + bvname))
@@ -185,7 +179,7 @@ func (nz *normalizer) convertComparison(node *ComparisonExpr) {
 			Value: bval.Value,
 		})
 	}
-	bvname := nz.newName()
+	bvname := nz.reserved.Next()
 	nz.bindVars[bvname] = bvals
 	// Modify RHS to be a list bindvar.
 	node.Right = ListArg(append([]byte("::"), bvname...))
@@ -211,17 +205,6 @@ func (nz *normalizer) sqlToBindvar(node SQLNode) *querypb.BindVariable {
 		return sqltypes.ValueBindVariable(v)
 	}
 	return nil
-}
-
-func (nz *normalizer) newName() string {
-	for {
-		newName := nz.prefix + strconv.Itoa(nz.counter)
-		if _, ok := nz.reserved[newName]; !ok {
-			nz.reserved[newName] = struct{}{}
-			return newName
-		}
-		nz.counter++
-	}
 }
 
 // GetBindvars returns a map of the bind vars referenced in the statement.