… go/test/fuzzing/parser_fuzzer.go
… go/vt/sqlparser/analyzer_test.go
… go/vt/sqlparser/ast_rewriting_test.go
… go/vt/sqlparser/ast_test.go
… go/vt/sqlparser/normalizer_test.go
… go/vt/vtgate/engine/memory_sort_test.go
… go/vt/vtgate/planbuilder/bypass.go
… go/vt/vtgate/planbuilder/ddl.go
… go/vt/vtgate/planbuilder/delete.go
… go/vt/vtgate/planbuilder/dml.go
… go/vt/vtgate/planbuilder/explain.go
… go/vt/vtgate/planbuilder/expr.go
… go/vt/vtgate/planbuilder/expr_test.go
… go/vt/vtgate/planbuilder/fallback_planner.go
… go/vt/vtgate/planbuilder/fallback_planner_test.go
… go/vt/vtgate/planbuilder/from.go
… go/vt/vtgate/planbuilder/insert.go
… go/vt/vtgate/planbuilder/join.go
… go/vt/vtgate/planbuilder/jointab.go
… go/vt/vtgate/planbuilder/jointab_test.go
… go/vt/vtgate/planbuilder/locktables.go
… go/vt/vtgate/planbuilder/postprocess.go
… go/vt/vtgate/planbuilder/route.go
… go/vt/vtgate/planbuilder/route_planning.go
… go/vt/vtgate/planbuilder/select.go
… go/vt/vtgate/planbuilder/system_tables.go
… go/vt/vtgate/planbuilder/union.go
… go/vt/vtgate/planbuilder/update.go
… go/vt/vttablet/tabletmanager/vreplication/controller_plan.go
… go/vt/vttablet/tabletmanager/vreplication/table_plan_builder.go